refactor: moved common and currency config to root since being used globally
