+++ src/botPage/bot/TradeEngine/OpenContract.js
@@ -1,6 +1,6 @@
 import { api_base } from '@api-base';
 import { observer } from '@utilities/observer';
-import { roundBalance } from '../../common/tools';
+import { roundBalance } from '../helpers';
 import { doUntilDone } from '../tools';
 import { contractStatus, contractSettled, contract as broadcastContract } from '../broadcast';
 import { sell, openContractReceived } from './state/actions';
