+++ src/botPage/bot/TradeEngine/Total.js
@@ -1,6 +1,6 @@
 import { translate } from '@i18n';
 import { observer as globalObserver } from '@utilities/observer';
-import { roundBalance } from '../../common/tools';
+import { roundBalance } from '../helpers';
 import { info, notify } from '../broadcast';
 import { createError } from '../../common/error';
