+++ src/botPage/common/symbolApi/index.js
@@ -1,9 +1,9 @@
 import { api_base } from '@api-base';
 import { getTokenList, removeAllTokens } from '@storage';
 import { observer as globalObserver } from '@utilities/observer';
+import config from '@currency-config';
+import { getObjectValue } from '@utils';
 import ActiveSymbols from './activeSymbols';
-import config from '../const';
-import { getObjectValue } from '../../../common/utils/tools';
 
 let parsed_asset_index;
