+++ src/botPage/view/TradeInfoPanel/AnimateTrade.js
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import classNames from 'classnames';
 import { translate } from '@i18n';
 import { observer as globalObserver } from '@utilities/observer';
-import { roundBalance } from '../../common/tools';
+import { roundBalance } from '../../bot/helpers';
 import useIsMounted from '../../../common/hooks/isMounted';
 import { setIsBotRunning } from '../../../redux-store/ui-slice';
 import Stage from './components/Stage';
