+++ src/botPage/view/blockly/blocks/after_purchase/read_details.js
@@ -1,7 +1,7 @@
 // https://blockly-demo.appspot.com/static/demos/blockfactory/index.html#u8i287
 import { translate } from '@i18n';
+import config from '@currency-config';
 import { insideAfterPurchase } from '../../relationChecker';
-import config from '../../../../common/const';
 
 Blockly.Blocks.read_details = {
     init: function init() {
