+++ src/botPage/view/blockly/blocks/indicators/bba.js
@@ -1,6 +1,6 @@
 // https://blockly-demo.appspot.com/static/demos/blockfactory/index.html#3qghes
 import { translate } from '@i18n';
-import config from '../../../../common/const';
+import config from '@currency-config';
 import { expectValue } from '../shared';
 
 Blockly.Blocks.bba = {
