+++ src/botPage/view/blockly/blocks/tools/candle/ohlc_values_in_list.js
@@ -1,6 +1,6 @@
 // https://blockly-demo.appspot.com/static/demos/blockfactory/index.html#2jo335
 import { translate } from '@i18n';
-import config from '../../../../../common/const';
+import config from '@currency-config';
 import { expectValue } from '../../shared';
 
 Blockly.Blocks.ohlc_values_in_list = {
