+++ src/botPage/view/blockly/blocks/tools/candle/read_ohlc_obj.js
@@ -1,6 +1,6 @@
 // https://blockly-demo.appspot.com/static/demos/blockfactory/index.html#szwuog
 import { translate } from '@i18n';
-import config from '../../../../../common/const';
+import config from '@currency-config';
 import { expectValue } from '../../shared';
 
 Blockly.Blocks.read_ohlc_obj = {
