+++ src/botPage/view/blockly/blocks/trade/backwardCompatibility.js
@@ -1,5 +1,5 @@
 import { translate } from '@i18n';
-import config from '../../../../common/const';
+import config from '@currency-config';
 import { symbolApi } from '../../../shared';
 import { setInputList, marketDefPlaceHolders, marketToTradeOption } from './tools';
 import { duration, payout, prediction, barrierOffsetGenerator } from './components';
