+++ src/botPage/view/blockly/utils.js
@@ -2,8 +2,8 @@
 import { translate } from '@i18n';
 import { saveAs } from '@utils';
 import { observer as globalObserver } from '@utilities/observer';
+import config from '@currency-config';
 import { fieldGeneratorMapping } from './blocks/shared';
-import config from '../../common/const';
 import { TrackJSError } from '../logger';
 
 export const isMainBlock = blockType => config.mainBlocks.indexOf(blockType) >= 0;
