+++ src/botPage/view/react-components/ExportButton.js
@@ -1,6 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { exportButton as exportButtonStyle } from '../style';
+
+const exportButtonStyle = {
+    position: 'absolute',
+    left: '41em',
+    zIndex: 2,
+    padding: 0,
+    width: '2em',
+    height: '2em',
+};
 
 const ExportButton = ({ onClick, customStyle }) => (
     <button onClick={onClick} style={{ ...exportButtonStyle, ...customStyle }} className='icon-save' />
