+++ src/components/Dialogs/DialogComponent.js
