+++ src/components/Dialogs/IntegrationsDialog.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { translate } from '@i18n';
 import Dialog from './Dialog';
-import GoogleDriveIntegration from '../react-components/Integrations/GoogleDriveIntegration';
+import GoogleDriveIntegration from '../../botPage/view/react-components/Integrations/GoogleDriveIntegration';
 import * as style from '../style';
 
 const IntegrationsContent = () => (
