+++ src/components/Dialogs/TradingView.js
