+++ src/components/ToolBox/save/save.jsx
@@ -2,12 +2,12 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { translate } from '@i18n';
 import { observer as globalObserver } from '@utilities/observer';
+import google_drive_util from '@utilities/integrations/GoogleDrive';
 import LoadingButton from '../loading_button';
 import SAVE_LOAD_TYPE from '../common';
 import useIsMounted from '../../../common/hooks/isMounted';
 import { cleanBeforeExport } from '../../../botPage/view/blockly/utils';
-import * as style from '../../../botPage/view/style';
-import google_drive_util from '../../../common/integrations/GoogleDrive';
+import * as style from '../../style';
 
 const Save = ({ blockly, closeDialog, is_gd_logged_in }) => {
     const [is_loading, setLoading] = React.useState(false);
